Refactor authroized reques
