+++ pizza-time/src/containers/login/login.js
@@ -61,6 +61,8 @@ export default function Login(props) {
     }
   };
 
+  //TODO implement a sign in with google
+  //TODO added checks to make sure that the registered users registered on our back and
   const signInWithGoogle = async e => {
     e.preventDefault();
     try {
